Extract code from GCB
